implement AbstractReferenceUserInput class to replace Reference and ReferenceFreeText interface
