--- src/main/java/com/bwgjoseph/springmvcdynamicuserinput/userinput/Reference.java
@@ -1,6 +1,0 @@
-package com.bwgjoseph.springmvcdynamicuserinput.userinput;
-
-public interface Reference<T extends ReferenceEntity> extends UserInput  {
-    T getReferenceValue();
-    String getCollection();
-}
