--- src/main/java/com/bwgjoseph/springmvcdynamicuserinput/userinput/ReferenceFreeText.java
@@ -1,8 +1,0 @@
-package com.bwgjoseph.springmvcdynamicuserinput.userinput;
-
-import java.util.Optional;
-
-public interface ReferenceFreeText<T extends ReferenceEntity> extends UserInput  {
-    Optional<T> getReferenceValue();
-    String getCollection();
-}
